Add blacklisted video detection capabilities for YouTube content (#32)
It adds a check for blacklisted video on the current domain and a way to show the user the limitation before even trying to play it.

This blacklisted information is not exposed in Youtube regular API so we need an extra REST endpoint to do the check for us. This endpoint is not part of the Open Source project but can be configured at runtime thanks to the youtubeExtraVideosInfoUrl build parameter. MixTube keeps working without this extra REST API but is just not able to detect blacklisted videos upfront.